[refactor] auth page 폴더 구조 변경
