+++ src/Router.tsx
@@ -1,14 +1,14 @@
 import Layout from 'components/Common/Layout';
 import Activity from 'page/today/Activity';
-import Member from 'page/Member';
+import Member from 'page/auth/Member';
 import Profile from 'page/Profile';
-import Register from 'page/Register';
-import ResetEmailForm from 'page/ResetEmailForm';
-import ResetPassword from 'page/ResetPassword';
-import ResetPasswordForm from 'page/ResetPasswordForm';
+import Register from 'page/auth/Register';
+import ResetEmailForm from 'page/auth/ResetEmailForm';
+import ResetPassword from 'page/auth/ResetPassword';
+import ResetPasswordForm from 'page/auth/ResetPasswordForm';
 import Search from 'page/Search';
-import SignIn from 'page/SignIn';
-import SignUp from 'page/SignUp';
+import SignIn from 'page/auth/SignIn';
+import SignUp from 'page/auth/SignUp';
 import Temperatures from 'page/today/Temperatures';
 import Tpo from 'page/today/Tpo';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
