+++ src/page/auth/SignIn.tsx
@@ -1,9 +1,10 @@
 import Logo from 'components/Common/Logo';
 import React from 'react';
 import styled from 'styled-components';
-import { ReactComponent as EmailIconSvg } from '../styles/images/icons/email-icon.svg';
-import { ReactComponent as GoogleIconSvg } from '../styles/images/icons/google-icon.svg';
+import { ReactComponent as EmailIconSvg } from '../../styles/images/icons/email-icon.svg';
+import { ReactComponent as GoogleIconSvg } from '../../styles/images/icons/google-icon.svg';
 import AuthSelectButton from 'components/Auth/AuthSelectButton';
+import { Link } from 'react-router-dom';
 
 const Auth = styled.section`
   margin-top: 24px;
@@ -54,29 +55,30 @@ const AuthInfo = styled.div`
   color: ${({ theme }) => theme.colors.gray6};
 `;
 
-function Register() {
+function SignIn() {
   return (
     <Auth>
       <Logo />
-      <h2>오늘의 옷장 회원가입</h2>
+      <h2>오늘의 옷장 로그인</h2>
       <p>
-        이미 계정이 있으세요? <a href="/signin">로그인</a>
+        계정이 없으세요?
+        <a href="/signin/register">회원가입</a>
       </p>
       <AuthSocialLists>
         <AuthSelectButton
-          text="구글로 회원가입"
+          text="구글로 로그인"
           href="#"
           icon={<GoogleIconSvg />}
         />
 
         <AuthSelectButton
-          text="이메일로 회원가입"
-          href="/signup"
+          text="이메일로 로그인"
+          href="/signin/member"
           icon={<EmailIconSvg />}
         />
       </AuthSocialLists>
       <AuthFind>
-        <a href="#">아이디 찾기</a>
+        <Link to={'/reset'}>비밀번호 재설정</Link>
       </AuthFind>
 
       <AuthInfo>
@@ -88,4 +90,4 @@ function Register() {
   );
 }
 
-export default Register;
+export default SignIn;
