+++ src/page/auth/SignUp.tsx
@@ -1,10 +1,10 @@
 import Logo from 'components/Common/Logo';
 import React, { useEffect } from 'react';
 import styled from 'styled-components';
-import { ReactComponent as ValidationErrorSvg } from '../styles/images/icons/validation-error.svg';
+import { ReactComponent as ValidationErrorSvg } from '../../styles/images/icons/validation-error.svg';
 import { useForm } from 'react-hook-form';
 import { IEmailSignUpInputData } from 'types/Auth';
-import { emailRegex } from '../util/index';
+import { emailRegex } from '../../util/index';
 import Input from 'components/Auth/Input';
 import { RootState, useAppDispatch } from 'redux/store';
 import { authJoinByEmail } from 'redux/action/auth';
